(harmless) wrong cast

+++ gegl/buffer/gegl-sampler-nearest.c
@@ -72,8 +72,8 @@ gegl_sampler_nearest_get (GeglSampler     *self,
   gfloat *sampler_bptr;
 
   sampler_bptr = gegl_sampler_get_from_buffer (self,
-                                               (gdouble) floor ((double) x),
-                                               (gdouble) floor ((double) y),
+                                               (gint) floor ((double) x),
+                                               (gint) floor ((double) y),
                                                repeat_mode);
   babl_process (self->fish, sampler_bptr, output, 1);
 }
